move starterbricks into separate folders
