+++ src/analysis/analysisVisitors/selectorAnalysis.test.ts
@@ -22,7 +22,7 @@ import {
   menuItemFormStateFactory,
   triggerFormStateFactory,
 } from "@/testUtils/factories/pageEditorFactories";
-import type { AttachMode } from "@/starterBricks/menuItemExtension";
+import type { AttachMode } from "@/starterBricks/menuItem/menuItemExtension";
 import { HighlightEffect } from "@/bricks/effects/highlight";
 import brickRegistry from "@/bricks/registry";
 import { JQueryReader } from "@/bricks/transformers/jquery/JQueryReader";
