+++ src/contentScript/lifecycle.ts
@@ -34,7 +34,7 @@ import { type UUID } from "@/types/stringTypes";
 import { type RegistryId } from "@/types/registryTypes";
 import { RunReason } from "@/types/runtimeTypes";
 import { type ResolvedModComponent } from "@/types/modComponentTypes";
-import { type SidebarStarterBrickABC } from "@/starterBricks/sidebarExtension";
+import { type SidebarStarterBrickABC } from "@/starterBricks/sidebar/sidebarExtension";
 import {
   getReloadOnNextNavigate,
   setReloadOnNextNavigate,
@@ -46,7 +46,7 @@ import {
 } from "@/utils/promiseUtils";
 import { $safeFind } from "@/utils/domUtils";
 import { onContextInvalidated } from "webext-events";
-import { ContextMenuStarterBrickABC } from "@/starterBricks/contextMenu";
+import { ContextMenuStarterBrickABC } from "@/starterBricks/contextMenu/contextMenu";
 import { ReusableAbortController } from "abort-utils";
 import { isLoadedInIframe } from "@/utils/iframeUtils";
 import { renderPanelsIfVisible } from "@/contentScript/sidebarController";
