+++ src/starterBricks/factory.ts
@@ -15,14 +15,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { fromJS as deserializePanel } from "@/starterBricks/panelExtension";
-import { fromJS as deserializeMenuItem } from "@/starterBricks/menuItemExtension";
+import { fromJS as deserializePanel } from "@/starterBricks/panel/panelExtension";
+import { fromJS as deserializeMenuItem } from "@/starterBricks/menuItem/menuItemExtension";
 import { fromJS as deserializeTrigger } from "@/starterBricks/triggerExtension";
-import { fromJS as deserializeContextMenu } from "@/starterBricks/contextMenu";
-import { fromJS as deserializeSidebar } from "@/starterBricks/sidebarExtension";
-import { fromJS as deserializeQuickBar } from "@/starterBricks/quickBarExtension";
-import { fromJS as deserializeQuickBarProvider } from "@/starterBricks/quickBarProviderExtension";
-import { fromJS as deserializeTour } from "@/starterBricks/tourExtension";
+import { fromJS as deserializeContextMenu } from "@/starterBricks/contextMenu/contextMenu";
+import { fromJS as deserializeSidebar } from "@/starterBricks/sidebar/sidebarExtension";
+import { fromJS as deserializeQuickBar } from "@/starterBricks/quickBar/quickBarExtension";
+import { fromJS as deserializeQuickBarProvider } from "@/starterBricks/quickBarProvider/quickBarProviderExtension";
+import { fromJS as deserializeTour } from "@/starterBricks/tour/tourExtension";
 import { type StarterBrick } from "@/types/starterBrickTypes";
 import { type StarterBrickConfig } from "@/starterBricks/types";
 import { getPlatform } from "@/platform/platformContext";
