+++ src/starterBricks/quickBar/quickBarExtension.tsx
