+++ src/starterBricks/sidebar/sidebarExtension.ts
