+++ src/starterBricks/tour/tourActionIcon.tsx
